feat: split baseproject secret retrieval
after migration all terraform baseprojects to v1 we can assume that each
one, that wants to use gke_auth has a baseproject_k8s instance and
therefore a infos-k8s in vault. let GHA handle the logic instead of
using these insane workarounds necessary in shell with JSON an variable
expansion. this also avoids the error that is printed in the run summary
when infos_k8s is not found, which confused devs